backend/requirements.txt
fastapi
httpx
uvicorn
python-dotenv
google-generativeai
aiocache